Adding in security groups method
This function allows users to need only the name of the security group to add the lambda to the security group. As the ids for the security group are generated they're added to a list and actively added to the lambda config. This method can be called using the action "create" as as well as "update-config"