Fixing handler issue
While working on the test invocation methods I noticed that there was
an issue with the way the handler. The user now inputs the json
filename as the handler, then their main function name as the main
function. maestro puts those two together as the handler name. This is
tested and confirmed working.